requirements.txt
certifi==2019.11.28
Django==2.2
django-admin==2.0.0
django-excel-response2==2.0.8
django-filter==2.2.0
django-six==1.0.4
djangorestframework==3.11.0
Markdown==3.2
pytz==2019.3
screen==1.0.1
sqlparse==0.3.0
xlwt==1.3.0
